+++ lib/extension_tools.dart
@@ -6,4 +6,4 @@ library extension_tools;
 export 'src/string_extensions.dart';
 export 'src/date_time_extensions.dart';
 
-export 'src/test_resources/mock_mixin.dart';
+export 'test/mock_mixin.dart';